test: split frame and element metadata tests into different files
They test different pages, splitting them makes the tests more maintainable
and allows Playwright to run them in parallel.

RISDEV-4221
