+++ frontend/e2e/amending-law-metadata-editor-rahmen.spec.ts
@@ -1,145 +1,7 @@
-import { Page, expect, test } from "@playwright/test"
+import test, { Page, expect } from "@playwright/test"
 
 test.describe("navigate to page", () => {
-  test("navigate to affected document metadata editor", async ({ page }) => {
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents",
-    )
-
-    await page.getByRole("link", { name: "Metadaten bearbeiten" }).click()
-
-    // Only expect the URL to be somewhat equal to the following. The reason is that
-    // the page redirects to a subpage, so this test might be flaky otherwise depending
-    // on how fast the redirect is.
-    await expect(page).toHaveURL(
-      /.*\/amending-laws\/eli\/bund\/bgbl-1\/2023\/413\/2023-12-29\/1\/deu\/regelungstext-1\/affected-documents\/eli\/bund\/bgbl-1\/1990\/s2954\/2023-12-29\/1\/deu\/regelungstext-1\/edit.*/,
-    )
-  })
-
-  test("displays affected document title", async ({ page }) => {
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit",
-    )
-
-    await expect(page.getByText("BGBl. I 2023 Nr. 413")).toBeVisible()
-
-    await expect(
-      page.getByText(
-        "Gesetz zum ersten Teil der Reform des Nachrichtendienstrechts",
-      ),
-    ).toBeVisible()
-  })
-
-  test("shows an error when the amending law could not be loaded", async ({
-    page,
-  }) => {
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/0000/000/0000-00-00/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/",
-    )
-
-    await expect(page.getByText("404")).toBeVisible()
-  })
-
-  test("shows an error when the time boundaries could not be loaded", async ({
-    page,
-  }) => {
-    await page.route(/timeBoundaries/, (route) => {
-      route.abort()
-    })
-
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/1970-01-01",
-    )
-
-    await expect(
-      page.getByText("Die Zeitgrenzen konnten nicht geladen werden."),
-    ).toBeVisible()
-
-    await page.unrouteAll()
-  })
-
-  test("shows an error when an element page could not be loaded", async ({
-    page,
-  }) => {
-    await page.route(
-      /elements\/hauptteil-1_abschnitt-erster_para-6/,
-      async (route) => {
-        await route.abort()
-      },
-    )
-
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/1970-01-01/hauptteil-1_abschnitt-erster_para-6",
-    )
-
-    await expect(
-      page.getByText("Das Element konnte nicht geladen werden."),
-    ).toBeVisible()
-
-    await page.unrouteAll()
-  })
-})
-
-test.describe("sidebar navigation", () => {
-  test("shows the elements affected by this amending law", async ({ page }) => {
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/1970-01-01",
-    )
-
-    const nav = page.getByRole("complementary", { name: "Inhaltsverzeichnis" })
-
-    const links = nav.getByRole("link")
-    await expect(links).toHaveText(
-      [
-        "Rahmen",
-        "§ 6 Gegenseitige Unterrichtung der Verfassungsschutzbehörden",
-      ],
-      { useInnerText: true },
-    )
-
-    // Go to the other time boundary to check if the result is the same (should always
-    // show all affected elements independent from the time boundary).
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/2023-12-30",
-    )
-
-    // Wait again ...
-    await page.waitForResponse((response) =>
-      response.url().includes("/elements?type="),
-    )
-
-    // Expect to see the same result again
-    await expect(links).toHaveText(
-      [
-        "Rahmen",
-        "§ 6 Gegenseitige Unterrichtung der Verfassungsschutzbehörden",
-      ],
-      { useInnerText: true },
-    )
-  })
-
-  test("shows the available time boundaries in the document", async ({
-    page,
-  }) => {
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit",
-    )
-
-    const select = page.getByRole("combobox", { name: "Zeitgrenze" })
-    const options = select.getByRole("option")
-
-    // First entry selected by default
-    await expect(select).toHaveValue("1970-01-01")
-
-    // Time boundaries available as options
-    await expect(options).toHaveText(["01.01.1970", "30.12.2023"], {
-      useInnerText: true,
-    })
-  })
-
-  test("navigates to the selected time boundary of the whole document", async ({
-    page,
-  }) => {
+  test("navigates to the selected time boundary", async ({ page }) => {
     await page.goto(
       "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit",
     )
@@ -151,117 +13,11 @@ test.describe("sidebar navigation", () => {
     await expect(page).toHaveURL(
       "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/2023-12-30",
     )
-  })
-
-  test("restores the selected time boundary from the URL", async ({ page }) => {
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/2023-12-30",
-    )
-
-    await expect(
-      page.getByRole("combobox", { name: "Zeitgrenze" }),
-    ).toHaveValue("2023-12-30")
-  })
-
-  test("navigates to the selected time boundary of an element", async ({
-    page,
-  }) => {
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/",
-    )
-
-    await page
-      .getByRole("combobox", { name: "Zeitgrenze" })
-      .selectOption("2023-12-30")
-
-    const nav = page.getByRole("complementary", { name: "Inhaltsverzeichnis" })
-
-    await nav
-      .getByRole("link", {
-        name: "§ 6 Gegenseitige Unterrichtung der Verfassungsschutzbehörden",
-      })
-      .click()
-
-    await expect(page).toHaveURL(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/2023-12-30/hauptteil-1_abschnitt-erster_para-6",
-    )
-  })
-
-  test("shows an error when the elements could not be loaded", async ({
-    page,
-  }) => {
-    await page.route(
-      /elements\?type=article&type=conclusions&type=preamble&type=preface&amendedBy=eli/,
-      (route) => {
-        route.abort()
-      },
-    )
-
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/1970-01-01",
-    )
-
-    await expect(
-      page.getByText("Artikel konnten nicht geladen werden."),
-    ).toBeVisible()
-
-    await page.unrouteAll()
-  })
-
-  test("shows an empty state when no elements are found", async ({ page }) => {
-    await page.route(
-      /elements\?type=article&type=conclusions&type=preamble&type=preface&amendedBy=eli/,
-      async (route) => {
-        await route.fulfill({ status: 200, json: {} })
-      },
-    )
-
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/1970-01-01",
-    )
-
-    await expect(page.getByText("Keine Artikel gefunden.")).toBeVisible()
-
-    await page.unrouteAll()
-  })
-
-  test("does not render links when no time boundary is selected", async ({
-    page,
-  }) => {
-    page.route(/timeBoundaries/, async (route) => {
-      await route.fulfill({ json: [], status: 200 })
-    })
-
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit",
-    )
-
-    await expect(page.getByText("Keine Zeitgrenze ausgewählt.")).toBeVisible()
-
-    await expect(
-      page
-        .getByRole("complementary", { name: "Inhaltsverzeichnis" })
-        .getByRole("link"),
-    ).toHaveCount(0)
-  })
-
-  test("navigates between elements", async ({ page }) => {
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2024/108/2024-03-27/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/2009/s3366/2024-03-27/1/deu/regelungstext-1/edit/1934-10-16/hauptteil-1_para-2",
-    )
-
-    const heading = page.getByRole("heading", { level: 2 })
-    await expect(heading).toHaveText("§1")
-
-    await page.getByRole("link", { name: "§3a" }).click()
-    await expect(heading).toHaveText("§3a")
   })
 })
 
 test.describe("preview", () => {
-  test("displays the title and preview of the whole document", async ({
-    page,
-  }) => {
+  test("displays the title and preview", async ({ page }) => {
     await page.goto(
       "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/2023-12-30",
     )
@@ -276,9 +32,7 @@ test.describe("preview", () => {
     ).toBeVisible()
   })
 
-  test("shows the preview of the whole document at different time boundaries", async ({
-    page,
-  }) => {
+  test("shows the preview at different time boundaries", async ({ page }) => {
     await page.goto(
       "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/1970-01-01",
     )
@@ -296,50 +50,7 @@ test.describe("preview", () => {
     await expect(preview).toHaveText(/.*nach Ablauf von fünf Jahren.*/)
   })
 
-  test("displays the title and preview of an element", async ({ page }) => {
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/2023-12-30/hauptteil-1_abschnitt-erster_para-6",
-    )
-
-    await expect(
-      page.getByRole("heading", {
-        name: "§ 6 Gegenseitige Unterrichtung der Verfassungsschutzbehörden",
-        level: 2,
-      }),
-    ).toBeVisible()
-
-    const preview = page.getByRole("region", {
-      name: "Vorschau",
-    })
-    await expect(
-      preview.getByRole("heading", {
-        name: "§ 6 Gegenseitige Unterrichtung der Verfassungsschutzbehörden",
-        level: 3,
-      }),
-    ).toBeVisible()
-  })
-
-  test("shows the preview of an element at different time boundaries", async ({
-    page,
-  }) => {
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/1970-01-01/hauptteil-1_abschnitt-erster_para-6",
-    )
-
-    const preview = page.getByRole("region", { name: "Vorschau" })
-
-    await expect(preview).toHaveText(
-      /.*am Ende des Kalenderjahres, das dem Jahr der Protokollierung folgt.*/,
-    )
-
-    await page
-      .getByRole("combobox", { name: "Zeitgrenze" })
-      .selectOption("2023-12-30")
-
-    await expect(preview).toHaveText(/.*nach Ablauf von fünf Jahren.*/)
-  })
-
-  test("shows an error when the preview could not be loaded for the whole document", async ({
+  test("shows an error when the preview could not be loaded", async ({
     page,
   }) => {
     await page.route(
@@ -363,34 +74,9 @@ test.describe("preview", () => {
 
     await page.unrouteAll()
   })
-
-  test("shows an error when the preview could not be loaded for an element", async ({
-    page,
-  }) => {
-    await page.route(
-      /elements\/hauptteil-1_abschnitt-erster_para-6\?atIsoDate=/,
-      async (request) => {
-        await request.abort()
-      },
-    )
-
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/1970-01-01/hauptteil-1_abschnitt-erster_para-6",
-    )
-
-    const previewRegion = page.getByRole("region", {
-      name: "Vorschau",
-    })
-
-    await expect(
-      previewRegion.getByText("Die Vorschau konnte nicht geladen werden."),
-    ).toBeVisible()
-
-    await page.unrouteAll()
-  })
 })
 
-test.describe("XML preview", () => {
+test.describe("XML view", () => {
   test("displays the XML of the target law with metadata", async ({ page }) => {
     await page.goto(
       "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/1970-01-01",
@@ -463,9 +149,7 @@ test.describe("XML preview", () => {
     )
   })
 
-  test("shows an error when the XML could not be loaded for the whole document", async ({
-    page,
-  }) => {
+  test("shows an error when the XML could not be loaded", async ({ page }) => {
     await page.route(
       /\/norms\/eli\/bund\/bgbl-1\/1990\/s2954\/2023-12-29\/1\/deu\/regelungstext-1\?$/,
       async (request) => {
@@ -485,33 +169,10 @@ test.describe("XML preview", () => {
 
     await page.unrouteAll()
   })
-
-  test("shows an error when the XML could not be loaded for an element", async ({
-    page,
-  }) => {
-    await page.route(
-      /\/norms\/eli\/bund\/bgbl-1\/1990\/s2954\/2023-12-29\/1\/deu\/regelungstext-1\?$/,
-      async (request) => {
-        await request.abort()
-      },
-    )
-
-    await page.goto(
-      "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/1970-01-01/hauptteil-1_abschnitt-erster_para-6",
-    )
-
-    await page.getByRole("tab", { name: "XML" }).click()
-
-    await expect(
-      page.getByText("Die XML-Ansicht konnte nicht geladen werden."),
-    ).toBeVisible()
-
-    await page.unrouteAll()
-  })
 })
 
 test.describe("metadata reading", () => {
-  test("displays metadata on the frame", async ({ page }) => {
+  test("displays metadata", async ({ page }) => {
     await page.goto(
       "/amending-laws/eli/bund/bgbl-1/2023/413/2023-12-29/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/1990/s2954/2023-12-29/1/deu/regelungstext-1/edit/1970-01-01",
     )
@@ -548,9 +209,7 @@ test.describe("metadata reading", () => {
     )
   })
 
-  test("displays metadata on the frame at different time boundaries", async ({
-    page,
-  }) => {
+  test("displays metadata at different time boundaries", async ({ page }) => {
     await page.goto(
       "/amending-laws/eli/bund/bgbl-1/2024/108/2024-03-27/1/deu/regelungstext-1/affected-documents/eli/bund/bgbl-1/2009/s3366/2024-03-27/1/deu/regelungstext-1/edit/1934-10-16",
     )
@@ -680,7 +339,7 @@ test.describe("metadata reading", () => {
     )
   })
 
-  test("displays an error if the data could not be loaded for the whole document", async ({
+  test("displays an error if the data could not be loaded", async ({
     page,
   }) => {
     await page.route(/\/proprietary\/2023-12-30$/, (request) => {
@@ -1058,7 +717,7 @@ test.describe("metadata editing", () => {
     await page.unrouteAll()
   })
 
-  test("displays a success message when the data has been saved for the whole document", async ({
+  test("displays a success message when the data has been saved", async ({
     page,
   }) => {
     await page.goto(
@@ -1074,9 +733,7 @@ test.describe("metadata editing", () => {
     await page.unrouteAll()
   })
 
-  test("displays an error if the data could not be saved for the whole document", async ({
-    page,
-  }) => {
+  test("displays an error if the data could not be saved", async ({ page }) => {
     await page.route(/\/proprietary\/2023-12-30$/, (route) => {
       if (route.request().method() === "PUT") route.abort("failed")
       else route.continue()
